ログアウト機能実装＋リファ
