+++ front/src/app/(auth)/expense/page.tsx
@@ -158,11 +158,7 @@ const fetchData = async () => {
 
   return (
     <main>
-      <Head>
-        <title>家計簿アプリ</title>
-        <meta name="description" content="家計簿アプリです。" />
-      </Head>
-      <div className="container mx-auto p-10">
+      <div className="container mx-auto p-10 max-w-screen-2xl">
         <h1 className="text-2xl font-bold mb-4">家計簿一覧</h1>
         <form onSubmit={handleSubmit}>
           <div className="flex flex-col flex-wrap justify-between gap-3 md:flex-row">
