+++ local/default.conf
@@ -15,6 +15,11 @@ server {
         proxy_pass http://192.168.99.1:3000/; 
 
     }
+
+    location = / {
+        rewrite ^ /login redirect;
+    }
+
     location ~ ^/(expenses|auth|check-token) {
         proxy_pass http://192.168.99.1:8080;
     }
