+++ web/default.conf
@@ -14,6 +14,10 @@ server {
         proxy_pass  http://localhost:3000;
     }
 
+    location = / {
+        rewrite ^ /login redirect;
+    }
+
     location ~ ^/(expenses|auth|check-token) {
         proxy_pass http://localhost:8080;
     }
